refactor: order service -> user service (docs: comment)
